move class into components folder
