+++ shared/components/EventList.js
